pyproject.toml
[project]
name = "django-admin-style-linker"
version = "0.0.1"
description = "A Django admin mixin to link MonacoEditor CSS and TinyMCE HTML fields for a live preview."
authors = [
    { name = "Leonid Yesypov", email = "leonidyesipov@gmail.com" },
]
license = "MIT"
readme = "README.md"
keywords = ["django", "admin", "css", "preview", "tinymce", "monaco"]

classifiers = [
    "Environment :: Web Environment",
    "Operating System :: OS Independent",
    "Intended Audience :: Developers",
    "Natural Language :: English",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Framework :: Django :: 5.0",
    "Framework :: Django :: 5.1",
    "Framework :: Django :: 5.2",
]
requires-python = ">=3.11,<4"
dependencies = [
    "Django>=5.0",
    "django-admin-monaco-editor>=1.1.5",
    "django-tinymce>=4.0.0",
]

[project.urls]
"Homepage" = "https://github.com/john-psina/django-admin-style-linker"
"Repository" = "https://github.com/john-psina/django-admin-style-linker"
"Bug Tracker" = "https://github.com/john-psina/django-admin-style-linker/issues"
"Documentation" = "https://github.com/john-psina/django-admin-style-linker"



[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["admin_style_linker*"]

[tool.semantic_release]
version_toml = ["pyproject.toml:project.version"]
version_variables = ["admin_style_linker/__init__.py:__version__"]
branch = "main"
upload_to_vcs_release = true
upload_to_pypi = false
build_command = "pip install build && python -m build"

[tool.semantic_release.remote]
type = "github"

[tool.semantic_release.commit_parser_options]
allowed_tags = ["build", "chore", "ci", "docs", "feat", "fix", "perf", "style", "refactor", "test"]
major_tags = ["breaking"]
minor_tags = ["feat"]
patch_tags = ["fix", "perf"]

[tool.semantic_release.changelog]
exclude_commit_patterns = []

[tool.semantic_release.changelog.environment]
keep_trailing_newline = true

[tool.isort]
profile = "black"
line_length = 120
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.black]
line-length = 120
target-version = ['py312']
include = '\.pyi?$'
